Move SelectQueryable to Dialect module
