+++ sqldelight-compiler/dialect/build.gradle
@@ -11,6 +11,7 @@ dependencies {
   api libs.kotlinPoet
 
   compileOnly libs.intellij.coreImpl
+  compileOnly libs.intellij.utilEx
 }
 
 apply from: "$rootDir/gradle/gradle-mvn-push.gradle"
