+++ sqldelight-compiler/dialect/src/main/kotlin/app/cash/sqldelight/dialect/api/PragmaWithResults.kt
@@ -1,6 +1,5 @@
-package app.cash.sqldelight.core.compiler.model
+package app.cash.sqldelight.dialect.api
 
-import app.cash.sqldelight.dialect.api.QueryWithResults
 import com.alecstrong.sql.psi.core.psi.NamedElement
 import com.alecstrong.sql.psi.core.psi.QueryElement
 import com.alecstrong.sql.psi.core.psi.QueryElement.QueryResult
@@ -9,13 +8,13 @@ import com.alecstrong.sql.psi.core.psi.SqlPragmaStmt
 import com.alecstrong.sql.psi.core.psi.impl.SqlPragmaNameImpl
 import com.intellij.lang.ASTNode
 
-class PragmaWithResults(private val pragmaStmt: SqlPragmaStmt) : QueryWithResults {
+class PragmaWithResults(pragmaStmt: SqlPragmaStmt) : QueryWithResults {
   override var statement: SqlAnnotatedElement = pragmaStmt
   override val select: QueryElement = pragmaStmt.pragmaName as SqlDelightPragmaName
   override val pureTable: NamedElement? = null
 }
 
-internal class SqlDelightPragmaName(node: ASTNode?) : SqlPragmaNameImpl(node), QueryElement {
+class SqlDelightPragmaName(node: ASTNode?) : SqlPragmaNameImpl(node), QueryElement {
   override fun queryExposed() = listOf(
     QueryResult(
       column = this,
