+++ sqldelight-compiler/dialect/src/main/kotlin/app/cash/sqldelight/dialect/api/SelectQueryable.kt
@@ -1,7 +1,5 @@
-package app.cash.sqldelight.core.compiler.model
+package app.cash.sqldelight.dialect.api
 
-import app.cash.sqldelight.core.lang.util.parentOfType
-import app.cash.sqldelight.dialect.api.QueryWithResults
 import com.alecstrong.sql.psi.core.psi.NamedElement
 import com.alecstrong.sql.psi.core.psi.QueryElement.QueryColumn
 import com.alecstrong.sql.psi.core.psi.Queryable
@@ -11,6 +9,7 @@ import com.alecstrong.sql.psi.core.psi.SqlCreateViewStmt
 import com.alecstrong.sql.psi.core.psi.SqlCteTableName
 import com.alecstrong.sql.psi.core.psi.SqlViewName
 import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.psi.util.parentOfType
 
 class SelectQueryable(
   override val select: SqlCompoundSelectStmt,
