+++ sqldelight-compiler/src/main/kotlin/app/cash/sqldelight/core/lang/ParserUtil.kt
@@ -2,10 +2,10 @@ package app.cash.sqldelight.core.lang
 
 import app.cash.sqldelight.core.SqlDelightProjectService
 import app.cash.sqldelight.core.SqldelightParserUtil
-import app.cash.sqldelight.core.compiler.model.SqlDelightPragmaName
 import app.cash.sqldelight.core.lang.psi.FunctionExprMixin
 import app.cash.sqldelight.dialect.api.SqlDelightDialect
 import app.cash.sqldelight.dialect.api.SqlDelightModule
+import app.cash.sqldelight.dialect.api.SqlDelightPragmaName
 import com.alecstrong.sql.psi.core.SqlParserUtil
 import com.alecstrong.sql.psi.core.psi.SqlTypes
 import com.intellij.openapi.project.Project
