+++ sqldelight-compiler/src/main/kotlin/app/cash/sqldelight/core/lang/SqlDelightQueriesFile.kt
@@ -26,7 +26,6 @@ import app.cash.sqldelight.core.compiler.model.NamedMutator.Update
 import app.cash.sqldelight.core.compiler.model.NamedQuery
 import app.cash.sqldelight.core.lang.psi.StmtIdentifierMixin
 import app.cash.sqldelight.core.lang.util.argumentType
-import app.cash.sqldelight.core.lang.util.parentOfTypeOrNull
 import app.cash.sqldelight.core.lang.util.table
 import app.cash.sqldelight.core.psi.SqlDelightStmtList
 import com.alecstrong.sql.psi.core.SqlAnnotationHolder
@@ -38,6 +37,7 @@ import com.intellij.psi.FileViewProvider
 import com.intellij.psi.PsiDirectory
 import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.psi.util.parentOfType
 
 class SqlDelightQueriesFile(
   viewProvider: FileViewProvider,
@@ -114,7 +114,7 @@ class SqlDelightQueriesFile(
   internal val requiredAdapters by lazy {
     val binders = PsiTreeUtil.findChildrenOfType(this, SqlBindExpr::class.java)
     val argumentAdapters = binders.mapNotNull {
-      it.parentOfTypeOrNull<SqlInsertStmt>()?.let {
+      it.parentOfType<SqlInsertStmt>()?.let {
         if (it.acceptsTableInterface() && it.table.needsAdapters()) return@mapNotNull it.table.adapterProperty()
       }
       typeResolver.argumentType(it).parentAdapter()
